+++ .eslintrc.cjs
@@ -9,4 +9,5 @@ module.exports = {
   rules: {
     "@typescript-eslint/ban-ts-comment": "off",
   },
+  ignorePatterns: ["dist/", "node_modules/"],
 };